+++ src/components/TodoCounter/TodoCounter.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './TodoCounter.css'
 function TodoCounter({total, completed}){
     return(
-        <h2 className='TodoCounter'>Done {completed} of {total} TODOs</h2>
+        <h2 className='TodoCounter'>Done {total} of {completed} TODOs</h2>
     )
 }
