+++ src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-
-import App from './App';
-
+// import App from './App';
+import App from './components/App/App'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
